Expr

Documentation

data Expr

These are Expressions used in Command

Constructors

Add Expr Expr
Addition

Sub Expr Expr
Subtration

Mul Expr Expr
Multiplication

Div Expr Expr
Division

Val Lit
Single number

Ident Name
Single identifier

Abs Expr
Absolute value

Mod Expr Expr
Modulo

Power Expr Expr
Power

data Command

These are the REPL commands

Constructors

Set Name Expr
Setting expression to the variable

Eval Expr
Expression evaluation

Print Expr
Printing the expression

Loop Int String
Looping, the number of times and expression as a string

FunctionInit Name String
Initialize the function, function name, expression as a string

FunctionCall Name
Calling function

Simplify Expr
Simplify an expression

Load Name
Load a file and execute its commands

History Int
Index into history and execute selected command

Help
Display how to use certain commands

Quit
Quit the application

eval

Arguments

:: Tree (Name, Lit)
Variable name to value mapping

-> Expr
Expression to evaluate

-> Either Msg Lit
Error message or numeric result

Evaluation function evaluates a given expression and produces the result

val

Arguments

:: (Lit -> Lit -> Lit)
Operator function

-> Tree (Name, Lit)
Tree of variables

-> Expr
Left expression

-> Expr
Right expression

-> Either Msg Lit
Error message or numeric result

Apply an operator to the evaluation of the left and right expressions

pCommand :: Parser Command

Parse commands

pExpr :: Parser Expr

Parse expression

pFactor :: Parser Expr

Parse factors

pTerm :: Parser Expr

Parse terms