REPL

Documentation

data State

Maintain state of program

Constructors

State

Fields

vars :: Tree (Name, Float)

funcs :: Tree (Name, [String])

numCalcs :: Int

history :: [Command]

commands :: [String]

initState :: State

Initialize state to default values

updateTreeVars :: Ord a => a -> b -> Tree (a, b) -> Tree (a, b)

Add variable to tree of variables

addVar :: Name -> Float -> State -> State

Update variable key-value pairs list

addHistory :: State -> Command -> State

Add a command to the command history in the state

getHistory :: State -> Int -> Either String Command

Get the most recent command from the history

addCommands :: State -> [String] -> State

Add commands to the list of commands to be executed

hasCommands :: State -> Bool

Returns true if the list of commands isn't empty

removeCommand :: State -> State

Removes top-most command from list

load :: [String] -> State -> IO ()

Handles the errors of loading a file

exec :: [String] -> State -> IO ()

Execute a program command

readLine :: State -> IO String

Get next command in the list of commands in state, otherwise read command from console

process :: State -> Command -> IO ()

Set a variable and update the state

repl :: State -> IO ()

Read, Eval, Print Loop ^ This reads and parses the input using the pCommand parser, and calls ^ process to process the command. ^ process will call repl when done, so the system loops.