Expr

Documentation

data Expr

These are Expressions used in Command

Constructors

Add Expr Expr
Addition

Sub Expr Expr
Subtration

Mul Expr Expr
Multiplication

Div Expr Expr
Division

Val Float
Single number

Name Name
Single identifier

Err String
Error message

Abs Expr
Absolute value

Mod Expr Expr
Modulo

Power Expr Expr
Power

data Command

These are the REPL commands

Constructors

Set Name Expr
Setting expression to the variable

Eval Expr
Expression evaluation

Print String
Printing the expression

Loop Int String
Looping, the number of times and expression as a string

FunctionInit Name String
Initialize the function, function name, expression as a string

FunctionCall Name
Calling function

Simplify Expr

eval

Arguments

:: Tree (Name, Float)
Variable name to value mapping

-> Expr
Expression to evaluate

-> Either String Float
Result

Evaluation function evaluates a given expression and produces the result

val

Arguments

:: (Float -> Float -> Float)
Operator function

-> Tree (Name, Float)
Tree of variables

-> Expr
Left expression

-> Expr
Right expression

-> Either String Float
Result

Apply an operator to the evalution of the left and right expressions

pCommand :: Parser Command

Parse commands

pExpr :: Parser Expr

Parse expression

pFactor :: Parser Expr

Parse factors

pTerm :: Parser Expr

Parse terms